+++ libs/angular/src/v-angular/modal/dialog/dialog.stories.ts
@@ -1,0 +1,47 @@
+import { CommonModule } from '@angular/common'
+
+import { NggCoreWrapperModule } from '@sebgroup/green-angular/src/lib/shared'
+
+import '../modal.globals'
+
+import { CUSTOM_ELEMENTS_SCHEMA, importProvidersFrom } from '@angular/core'
+import {
+  applicationConfig,
+  Meta,
+  moduleMetadata,
+  StoryFn,
+} from '@storybook/angular'
+
+import { NggvI18nModule } from '../../i18n/i18n.module'
+import { NggvDialogComponent } from './dialog.component'
+
+export default {
+  title: 'V-Angular/Dialog',
+  component: NggvDialogComponent,
+  decorators: [
+    applicationConfig({
+      providers: [importProvidersFrom(NggvI18nModule)],
+    }),
+    moduleMetadata({
+      imports: [CommonModule, NggvI18nModule, NggCoreWrapperModule],
+      schemas: [CUSTOM_ELEMENTS_SCHEMA],
+    }),
+  ],
+} as Meta
+
+const Template: StoryFn<NggvDialogComponent> = (args: any) => ({
+  props: args,
+})
+
+export const Primary = Template.bind({})
+Primary.args = {
+  initiallyShown: true,
+  closeButtonAriaLabel: 'Close dialog',
+  buttons: {
+    negative: 'button_cancel',
+    neutral: 'button_apply',
+    positive: 'button_save',
+  },
+  content:
+    'You can supply the content seen here either through the <code>[content]="string"</code> property or</br>by passing children between the opening and closing tags <code>&lt;c-dialog&gt; ...children &lt;/c-dialog&gt;</code>',
+}