+++ libs/angular/src/v-angular/radio/radio.component.spec.ts
@@ -5,32 +5,45 @@ import {
   NgControl,
   ReactiveFormsModule,
 } from '@angular/forms'
+import { TranslocoTestingModule } from '@jsverse/transloco'
 
-import { TranslocoModule } from '@jsverse/transloco';
-import { NgvI18nTestModule } from '@sebgroup/green-angular/src/v-angular/i18n'
+import { NggvI18nTestModule } from '@sebgroup/green-angular/src/v-angular/i18n'
+import en from '../i18n/i18n.json'
 
 import '../core/core.globals'
 
-import { NgvRadioComponent } from './radio.component'
+import { NggvRadioComponent } from './radio.component'
 
-describe('[NgvCore]', () => {
+describe('[NggvCore]', () => {
   // ----------------------------------------------------------------------------
   // RadioComponent - constructor()
   // ----------------------------------------------------------------------------
   describe('RadioComponent - constructor()', () => {
-    let component: NgvRadioComponent
-    let fixture: ComponentFixture<NgvRadioComponent>
+    let component: NggvRadioComponent
+    let fixture: ComponentFixture<NggvRadioComponent>
 
     beforeEach(waitForAsync(() => {
       TestBed.configureTestingModule({
-        declarations: [NgvRadioComponent],
+        declarations: [NggvRadioComponent],
         providers: [{ provide: NgControl, useValue: new FormControl() }],
-        imports: [FormsModule, ReactiveFormsModule, NgvI18nTestModule, TranslocoModule,],
+        imports: [
+          FormsModule,
+          ReactiveFormsModule,
+          NggvI18nTestModule,
+          TranslocoTestingModule.forRoot({
+            langs: { en },
+            translocoConfig: {
+              availableLangs: ['en'],
+              defaultLang: 'en',
+            },
+            preloadLangs: true,
+          }),
+        ],
       }).compileComponents()
     }))
 
     beforeEach(() => {
-      fixture = TestBed.createComponent(NgvRadioComponent)
+      fixture = TestBed.createComponent(NggvRadioComponent)
       component = fixture.componentInstance
       fixture.detectChanges()
     })