+++ libs/angular/src/v-angular/tooltip/tooltip.directive.ts
@@ -13,6 +13,8 @@ import {
   SimpleChanges,
 } from '@angular/core'
 
+import { tooltipArrowStyles, tooltipBoxStyles } from './tooltip.styles'
+
 export type Placement = 'top' | 'right' | 'bottom' | 'left'
 
 type Position = {
@@ -31,7 +33,7 @@ type Position = {
 @Directive({
   selector: '[nggvTooltip]',
 })
-export class NgvTooltipDirective
+export class NggvTooltipDirective
   implements AfterViewInit, OnChanges, OnDestroy
 {
   /** The text that will be shown in the tooltip. */
@@ -48,6 +50,8 @@ export class NgvTooltipDirective
   @Input() offset = 10
   /** How frequently the tooltip will be re-rendered when the page size changes. */
   @Input() resizeThrottle = 50
+  /** Id of tooltip element. */
+  @Input() tooltipId?: string
   /** Numeric max-width for tooltip. */
   @Input() maxWidth = 343
 
@@ -95,18 +99,18 @@ export class NgvTooltipDirective
 
   ngOnChanges(changes: SimpleChanges): void {
     if (
-      NgvTooltipDirective.shouldUpdate(changes.nggvTooltip) ||
-      NgvTooltipDirective.shouldUpdate(changes.placement) ||
-      NgvTooltipDirective.shouldUpdate(changes.thook)
+      NggvTooltipDirective.shouldUpdate(changes.nggvTooltip) ||
+      NggvTooltipDirective.shouldUpdate(changes.placement) ||
+      NggvTooltipDirective.shouldUpdate(changes.thook)
     ) {
       this.shown ? this.show(true) : this.hide(true)
     }
 
-    if (NgvTooltipDirective.shouldUpdate(changes.shown)) {
+    if (NggvTooltipDirective.shouldUpdate(changes.shown)) {
       this.shown ? this.show() : this.hide()
     }
 
-    if (NgvTooltipDirective.shouldUpdate(changes.offset)) {
+    if (NggvTooltipDirective.shouldUpdate(changes.offset)) {
       this.updatePosition()
     }
   }
@@ -117,12 +121,15 @@ export class NgvTooltipDirective
 
   /** @internal */
   @HostListener('mouseenter')
+  @HostListener('focus')
   onMouseEnter() {
     this.show()
   }
 
   /** @internal */
   @HostListener('mouseleave')
+  @HostListener('blur')
+  @HostListener('keyup.escape')
   onMouseLeave() {
     this.hide()
   }
@@ -163,8 +170,9 @@ export class NgvTooltipDirective
    */
   hide(destroy = false) {
     if (!this.tooltipElement) return
-    if (this.parentElement.contains(this.tooltipElement))
+    if (this.parentElement.contains(this.tooltipElement)) {
       this.renderer.removeChild(this.parentElement, this.tooltipElement)
+    }
     if (destroy) this.destroy()
     this.shown = false
     this.nggvHide.emit(this.tooltipElement)
@@ -179,10 +187,14 @@ export class NgvTooltipDirective
     this.tooltipElement = this.renderer.createElement('div')
     this.renderer.addClass(this.tooltipElement, 'gds-tooltip')
     this.renderer.setAttribute(this.tooltipElement, 'data-thook', this.thook)
-    this.renderer.setStyle(this.tooltipElement, 'position', 'absolute')
-    this.renderer.setStyle(this.tooltipElement, 'z-index', '1040')
-    this.renderer.setStyle(this.tooltipElement, 'border-radius', '.25rem')
-    this.renderer.setStyle(this.tooltipElement, 'padding', '.5rem 1rem')
+    this.renderer.setAttribute(this.tooltipElement, 'role', 'tooltip')
+    if (this.tooltipId) {
+      this.renderer.setAttribute(this.tooltipElement, 'id', this.tooltipId)
+    }
+    // set styling
+    Array.from(tooltipBoxStyles.entries()).forEach(([style, value]) => {
+      this.renderer.setStyle(this.tooltipElement, style, value)
+    })
     const relativeMaxWidth = this.pxToRem(this.maxWidth)
     this.renderer.setStyle(this.tooltipElement, 'max-width', relativeMaxWidth)
     this.renderer.appendChild(
@@ -332,6 +344,12 @@ export class NgvTooltipDirective
       const position = this.pxToRem(value)
       this.renderer.setStyle(this.arrowElement, prop, position)
     })
+
+    Array.from(tooltipArrowStyles[this.placement].entries()).forEach(
+      ([style, value]) => {
+        this.renderer.setStyle(this.arrowElement, style, value)
+      },
+    )
   }
 
   private pxToRem(value: number): string {